\name{vbv-package}
\alias{vbv-package}
\alias{vbv}
\docType{package}
\title{VBV -- Time series decomposition with VBV (Verallgemeinertes Berliner Verfahren)}
}
\description{
vbv implements the VBV as described in the upcoming article
    "Decomposition of Time Series using the Generalised Berlin Method
    (VBV)" by Hebbel and Steuer, to be published in Jan Beran,
    Yuanhua Feng, Hartmut Hebbel (Eds.): Empirical Economic and
    Financial Research - Theory, Methods and Practice, Festschrift in
    Honour of Prof. Siegfried Heiler. Series: Advanced Studies in
    Theoretical and Applied Econometrics. Springer 2014, p. 9-40.
    VBV is a method to decompose a time series using
    penalized splines. Both methods, static estimation and moving
    estimation are implemented.
}
\details{
\tabular{ll}{
Package: \tab vbv\cr
Type: \tab Package\cr
Version: \tab 0.5 \cr
Date: \tab 2013-12-16\cr
License: GPL-3 or later
}

}
\author{
Hartmut Hebbel ( hartmut.hebbel@hsu-hh.de ) \cr
Detlef Steuer  (steuer@hsu-hh.de)

Maintainer: steuer@hsu-hh.de

}
\references{
Yuanhua Feng, Hartmut Hebbel (Eds.): Empirical Economic and
    Financial Research - Theory, Methods and Practice, Festschrift in
    Honour of Prof. Siegfried Heiler. Series: Advanced Studies in
    Theoretical and Applied Econometrics. Springer 2014, p. 9-40.
}
\keyword{ package }
\keyword{ time series analysis }
\seealso{

}
\examples{

}
